+++ protocols/gossipsub/src/lib.rs
@@ -104,7 +104,7 @@ mod mcache;
 mod metrics;
 mod peer_score;
 mod protocol;
-mod rpc;
+mod queue;
 mod rpc_proto;
 mod subscription_filter;
 mod time_cache;